requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==3.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-data-seeder==0.2.0
django-filter==23.2
django-ipware==5.0.0
django-jazzmin==2.6.0
django-rest-authemail==2.1.7
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.6.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.1.1
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.29.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
zipp==3.11.0
